Fix: reply_count type mismatch in post queries
- Updated `PostResponse` struct in `backend/src/models/post.rs` to use `i64` for `reply_count` to match `COUNT(*)` return type (BIGINT) from PostgreSQL.
- Updated `backend/src/services/posts.rs` to cast `i32` reply count from `Post` entity to `i64` when creating response.
- Updated `backend/src/api/v4/posts.rs` to use `0::bigint` in SQL query for consistency.
- Added reproduction test `backend/tests/api_v4_posts.rs`.

+++ backend/src/api/v4/posts.rs
@@ -160,7 +160,7 @@ async fn get_post_thread(
     let replies: Vec<crate::models::post::PostResponse> = sqlx::query_as(
         r#"
         SELECT p.*, u.username, u.email, u.avatar_url,
-        0 as reply_count, NULL as last_reply_at
+        0::bigint as reply_count, NULL as last_reply_at
         FROM posts p
         LEFT JOIN users u ON p.user_id = u.id
         WHERE p.root_post_id = $1 AND p.deleted_at IS NULL
